requirements.txt
# Jira Data Center Attachment Analysis Tool - Dependencies

# Core HTTP and API
requests>=2.31.0
urllib3>=2.0.0

# Progress and CLI
tqdm>=4.66.0
click>=8.1.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data Processing
pandas>=2.0.0

# Export Formats
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Reporting
jinja2>=3.1.0

# Utilities
python-dateutil>=2.8.0
humanize>=4.9.0

# Development (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
